refactor: routes for ctpat inspection
