+++ src/app/dashboard/ctpat/[id]/inspection-list.tsx
